feat: wip - interaction refactor
