+++ tests/interactions/sex_test.py
@@ -1,0 +1,50 @@
+import pytest
+
+from titan.interactions import Sex
+from titan.agent import Relationship
+
+from conftest import FakeRandom
+
+@pytest.mark.unit
+def test_sex_transmission(make_model, make_agent):
+    model = make_model()
+    a = make_agent()
+    a.sex_role = "insertive"
+    p = make_agent()
+    p.sex_role = "receptive"
+    a.partners["Sex"] = set()
+    p.partners["Sex"] = set()
+    rel = Relationship(a, p, 10, bond_type="Sex")
+
+    a.hiv = True
+    a.hiv_time = model.time  # acute
+
+    rel.total_sex_acts = 0
+    model.params.calibration.acquisition = 10
+
+    model.params.calibration.acquisition = 5
+    model.params.calibration.sex.act = 10
+    model.run_random = FakeRandom(0.6)
+
+    # test partner becomes
+    Sex.interact(model, rel)
+    assert p.hiv
+
+
+@pytest.mark.unit
+def test_sex_transmission_do_nothing(make_model, make_agent):
+    model = make_model()
+    a = make_agent()
+    p = make_agent()
+    a.partners["Sex"] = set()
+    p.partners["Sex"] = set()
+    rel = Relationship(a, p, 10, bond_type="Sex")
+
+    with pytest.raises(ValueError):
+        Sex.interact(model, rel)
+
+    a.hiv = True
+    p.hiv = True
+
+    # test nothing happens
+    Sex.interact(model, rel)
