+++ tests/model_test.py
@@ -142,99 +142,6 @@ def test_get_transmission_probability(make_model, make_agent):
 
 
 @pytest.mark.unit
-def test_needle_transmission(make_model, make_agent):
-    model = make_model()
-    a = make_agent(race="white", DU="Inj", SO="HM")
-    p = make_agent(race="white", DU="Inj", SO="HF")
-
-    with pytest.raises(AssertionError):
-        model.injection_transmission(a, p)
-
-    a.hiv = True
-    a.hiv_time = model.time - 1  # acute
-
-    model.run_random = FakeRandom(-0.1)
-
-    model.injection_transmission(a, p)
-
-    assert p.hiv
-
-
-@pytest.mark.unit
-def test_sex_transmission(make_model, make_agent):
-    model = make_model()
-    a = make_agent()
-    a.sex_role = "insertive"
-    p = make_agent()
-    p.sex_role = "receptive"
-    a.partners["Sex"] = set()
-    p.partners["Sex"] = set()
-    rel = Relationship(a, p, 10, bond_type="Sex")
-
-    a.hiv = True
-    a.hiv_time = model.time  # acute
-
-    rel.total_sex_acts = 0
-    model.params.calibration.acquisition = 10
-
-    model.params.calibration.acquisition = 5
-    model.params.calibration.sex.act = 10
-    model.run_random = FakeRandom(0.6)
-
-    # test partner becomes
-    model.sex_transmission(rel)
-    assert p.hiv
-
-
-@pytest.mark.unit
-def test_sex_transmission_do_nothing(make_model, make_agent):
-    model = make_model()
-    a = make_agent()
-    p = make_agent()
-    a.partners["Sex"] = set()
-    p.partners["Sex"] = set()
-    rel = Relationship(a, p, 10, bond_type="Sex")
-
-    with pytest.raises(ValueError):
-        model.sex_transmission(rel)
-
-    a.hiv = True
-    p.hiv = True
-
-    # test nothing happens
-    model.sex_transmission(rel)
-
-
-@pytest.mark.unit
-def test_pca_interaction(make_model, make_agent):
-    model = make_model()
-    a = make_agent()
-    p = make_agent()
-    a.pca.opinion = 4
-    p.pca.opinion = 2
-    a.pca.awareness = True
-    a.partners["SexInj"] = set()
-    p.partners["SexInj"] = set()
-
-    model.run_random = FakeRandom(1.0)
-
-    model.pop.graph.add_edge(a, p)
-    model.pop.graph.add_edge(a, "edge")
-
-    model.time = 5
-
-    rel = Relationship(a, p, 10, bond_type="SexInj")
-    model.pca_interaction(rel, force=True)
-
-    assert p.pca.awareness
-
-    model.time += 1
-    model.pca_interaction(rel, force=True)
-
-    assert p.pca.opinion == 3
-
-
-@pytest.mark.unit
 def test_hiv_convert(make_model, make_agent):
     model = make_model()
     a = make_agent()
