+++ titan/interactions/__init__.py
@@ -1,0 +1,5 @@
+# makes src a package
+from .base_interaction import *
+from .pca import *
+from .sex import *
+from .injection import *
