+++ titan/interactions/base_interaction.py
@@ -1,0 +1,8 @@
+class BaseInteraction:
+
+    name: str = ""
+    """Name of interaction in the params file."""
+
+    @staticmethod
+    def interact(model, rel):
+        pass
