+++ titan/model.py
@@ -17,6 +17,7 @@ from . import probabilities as prob
 from . import utils
 from .parse_params import ObjMap
 from . import features
+from . import interactions
 
 
 class HIVModel:
@@ -73,6 +74,11 @@ class HIVModel:
             for feature in features.BaseFeature.__subclasses__()
             if self.params.features[feature.name]
         ]
+
+        self.interactions = {
+            interaction.name: interaction
+            for interaction in interactions.BaseInteraction.__subclasses__()
+        }
 
         # Set seed format. 0: pure random, else: fixed value
         self.run_seed = utils.get_check_rand_int(params.model.seed.run)
@@ -359,240 +365,12 @@ class HIVModel:
         if rel.agent1.incar.active or rel.agent2.incar.active:  # type: ignore[attr-defined]
             return False
 
-        # Agent 1 is HIV+, Agent 2 is not, Agent 2 is succept
-        if rel.agent1.hiv and not rel.agent2.hiv:
-            agent = rel.agent1
-            partner = rel.agent2
-        # If Agent 2 is HIV and Agent 1 is not, Agent 1 is succept
-        elif not rel.agent1.hiv and rel.agent2.hiv:
-            agent = rel.agent2
-            partner = rel.agent1
-        else:  # neither agent is HIV or both are
-            return False
+        agents_interacted = False
+        for interaction_type in interaction_types:
+            interaction = self.interactions[interaction_type]
+            agents_interacted = interaction.interact(self, rel) or agents_interacted
 
-        if self.params.features.pca and self.time >= self.params.pca.start_time:
-            if "pca" in interaction_types and rel.duration < rel.total_duration:
-                self.pca_interaction(rel)
-
-        if self.time >= self.params.hiv.start_time:
-            if "injection" in interaction_types:
-                self.injection_transmission(agent, partner)
-
-            if "sex" in interaction_types:
-                self.sex_transmission(rel)
-
-        return True
-
-    def pca_interaction(self, rel: "ag.Relationship", force=False):
-        """
-        Simulate peer change agent interactions. Knowledge if one agent is aware and one unaware,
-            opinion if one agent swaying the other.
-
-        args:
-            rel: The relationship PCA is happening in
-            force: Whether to force knowledge dissemination and influence
-        """
-
-        assert (
-            self.params.model.network.enable
-        ), "Network must be enabled for pca interactions"
-
-        def influence(agent, partner):
-            agent_init_opinion = agent.pca.opinion
-            partner_init_opinion = partner.pca.opinion
-            agent_influence = nx.closeness_centrality(self.pop.graph, agent)
-            partner_influence = nx.closeness_centrality(self.pop.graph, partner)
-
-            if agent_influence > partner_influence:
-                partner.pca.opinion = np.mean([agent.pca.opinion, partner.pca.opinion])
-            elif agent_influence < partner_influence:
-                agent.pca.opinion = np.mean([agent.pca.opinion, partner.pca.opinion])
-
-            if (
-                agent_init_opinion
-                < agent.location.params.pca.opinion.threshold
-                < agent.pca.opinion
-            ):
-                if self.run_random.random() < agent.location.params.pca.prep.prob:
-                    agent.prep.initiate(self, force=True)
-
-            elif (
-                partner_init_opinion
-                < partner.location.params.pca.opinion.threshold
-                < partner.pca.opinion
-            ):
-                if self.run_random.random() < partner.location.params.pca.prep.prob:
-                    partner.prep.initiate(partner, self, force=True)
-
-        def knowledge_dissemination(partner):
-            partner.pca.awareness = True
-            if (
-                partner.pca.opinion > partner.location.params.pca.opinion.threshold
-                and self.run_random.random() < partner.location.params.pca.prep.prob
-            ):
-                self.initiate(partner, force=True)
-
-        def knowledge_transmission_probability():
-            if rel.agent1.pca.awareness and rel.agent2.pca.awareness:
-                p = self.params.pca.opinion.transmission
-            else:
-                p = self.params.pca.knowledge.transmission
-
-            if num_acts == 1:
-                p_total_transmission = p
-            elif num_acts >= 1:
-                p_total_transmission = 1.0 - utils.binom_0(num_acts, p)
-            else:
-                p_total_transmission = 0
-
-            return p_total_transmission
-
-        acts_prob = self.run_random.random()
-        acts_bin = 0
-        current_p_value = 0.0
-
-        while acts_prob > current_p_value:
-            acts_bin += 1
-            current_p_value += self.params.partnership.pca.frequency[rel.bond_type][
-                acts_bin
-            ].prob
-        min = self.params.partnership.pca.frequency[rel.bond_type][acts_bin].min
-        max = self.params.partnership.pca.frequency[rel.bond_type][acts_bin].max
-        if min == max:
-            num_acts = min
-        else:
-            num_acts = self.run_random.randrange(min, max)
-
-        if num_acts < 1 and not force:
-            return
-        if rel.agent1.pca.awareness and not rel.agent2.pca.awareness:  # type: ignore[attr-defined]
-            if self.run_random.random() < knowledge_transmission_probability() or force:
-                knowledge_dissemination(rel.agent2)
-        elif not rel.agent1.pca.awareness and rel.agent2.pca.awareness:  # type: ignore[attr-defined]
-            if self.run_random.random() < knowledge_transmission_probability() or force:
-                knowledge_dissemination(rel.agent1)
-        elif rel.agent1.pca.awareness and rel.agent2.pca.awareness or force:  # type: ignore[attr-defined]
-            if self.run_random.random() < knowledge_transmission_probability() or force:
-                influence(rel.agent1, rel.agent2)
-
-    def injection_transmission(self, agent: "ag.Agent", partner: "ag.Agent"):
-        """
-        Simulate random transmission of HIV between two PWID agents through injection.
-
-        args:
-            agent: PWID agent with HIV
-            partner: PWID agent without HIV
-        """
-
-        assert agent.hiv
-        assert not partner.hiv
-        assert agent.drug_type == "Inj"
-        assert partner.drug_type == "Inj"
-
-        agent_race = agent.race
-        agent_sex_type = agent.sex_type
-
-        mean_num_acts = (
-            agent.location.params.demographics[agent_race][
-                agent_sex_type
-            ].injection.num_acts
-            * self.calibration.injection.act
-        )
-        share_acts = utils.poisson(mean_num_acts, self.np_random)
-
-        if (
-            agent.syringe_services.active or partner.syringe_services.active  # type: ignore[attr-defined]
-        ):  # syringe services program risk
-            p_unsafe_injection = features.SyringeServices.enrolled_risk
-        else:
-            # If sharing, minimum of 1 share act
-            if share_acts < 1:
-                share_acts = 1
-
-            p_unsafe_injection = agent.location.params.demographics[agent_race][
-                agent_sex_type
-            ].injection.unsafe_prob
-
-            if agent.hiv_dx or partner.hiv_dx:  # diagnosis risk reduction
-                p_unsafe_injection *= 1 - self.params.hiv.dx.risk_reduction.injection
-
-        for n in range(share_acts):
-            if self.run_random.random() > p_unsafe_injection:
-                share_acts -= 1
-
-        if share_acts >= 1:
-            p = self.get_transmission_probability("injection", agent, partner)
-
-            p_total_transmission: float
-            if share_acts == 1:
-                p_total_transmission = p
-            else:
-                p_total_transmission = 1.0 - utils.binom_0(share_acts, p)
-
-            if self.run_random.random() < p_total_transmission:
-                # if agent HIV+ partner becomes HIV+
-                self.hiv_convert(partner)
-
-    def sex_transmission(self, rel: "ag.Relationship"):
-        """
-        Simulate random transmission of HIV between two agents through Sex. One of the agents must have HIV.
-
-        args:
-            rel : Relationship
-        """
-
-        if rel.agent1.hiv:
-            agent = rel.agent1
-            partner = rel.agent2
-        elif rel.agent2.hiv:
-            agent = rel.agent2
-            partner = rel.agent1
-        else:
-            raise ValueError("rel must have an agent with HIV")
-
-        # HIV status of agent and partner
-        # Everything from here is only run if one of them is HIV+
-        if partner.hiv:
-            return
-
-        # unprotected sex probabilities for primary partnerships
-        mean_sex_acts = (
-            agent.get_number_of_sex_acts(self.np_random) * self.calibration.sex.act
-        )
-        total_sex_acts = utils.poisson(mean_sex_acts, self.np_random)
-
-        # Get condom usage
-        p_safe_sex = agent.location.params.demographics[agent.race][
-            agent.sex_type
-        ].safe_sex
-        # increase condom usage if diagnosed
-        if agent.hiv_dx or partner.hiv_dx:
-            # Calculate probability of safe sex given risk reduction
-            p_unsafe_sex = (1 - p_safe_sex) * (
-                1 - self.params.hiv.dx.risk_reduction.sex
-            )
-            p_safe_sex *= 1 - p_unsafe_sex
-
-        # Reduction of risk acts between partners for condom usage
-        unsafe_sex_acts = total_sex_acts
-        for n in range(unsafe_sex_acts):
-            if self.run_random.random() < p_safe_sex:
-                unsafe_sex_acts -= 1
-
-        if unsafe_sex_acts >= 1:
-            # agent is HIV+
-            rel.total_sex_acts += unsafe_sex_acts
-            p_per_act = self.get_transmission_probability("sex", agent, partner)
-
-            p_total_transmission: float
-            if unsafe_sex_acts == 1:
-                p_total_transmission = p_per_act
-            else:
-                p_total_transmission = 1.0 - utils.binom_0(unsafe_sex_acts, p_per_act)
-
-            if self.run_random.random() < p_total_transmission:
-                # if agent HIV+ partner becomes HIV+
-                self.hiv_convert(partner)
+        return agents_interacted
 
     def get_transmission_probability(self, interaction: str, agent, partner) -> float:
         """
